requirements.txt
dash==2.14.1
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
huggingface-hub==0.20.1
python-dotenv==1.1.1
datasets==2.16.1
pillow==10.1.0
